Update README.md
# python-interview-coding
This repository contains the solution of average coding problems asked in the SDE level 1 and 2.
All the solutions are in python.

Questions- 

q1. Given an array consisting many inner arrays, flatten the array into one: example: input:[[6,4,7,[9,5,4,[2,4,8]]],[2,2,7],[9,0,7,[9,3,1,8,5]]] 
output: [6,4,7,9,5,4,2,4,8,2,2,7,9,0,7,9,3,1,8,5]

q2. 21 sticks problem: first the opponent will pickup the stick- minimumsticks = 1, maximum = 4, then you will pickup the stick.
the one picking up the last stick will be loser
write a code that will ensure that you will never loose

q3. Check if the string contains balance parenthesis or not

q4. Given an array of distinct elements and a number x, find if there is a pair with product equal to x.
and print all the pairs

q5. Given a string ‘str’ of digits and an integer ‘n’,
build the lowest possible number by removing ‘n’ digits from the string and not changing the order of input digits.
str = "4325043" , n=3
output = "2043"

q6. write a function that returns the Run Length Encoded string for the input string.
For example, if the input string is “aaabbccddd”, then the function should return “3a2b2c3d”

q7. You are given an array of integers. Return an array of the same size where the element at each index 
is the product of all the elements in the original array except for the element at that index. 
For example, an input of [1, 2, 3, 4, 5] should return [120, 60, 40, 30, 24]. You cannot use division in this problem.